Slash Command Menu works!
